Support for BulkData with 64-bit sizes. This is implemented via an additional BULKDATA flag, rather than with versioning, as BulkData is often loaded from unversioned archives, e.g. from DDC via FMemoryReader. This addresses the issue of enormous meshes >2Gb being unimportable (since source model data is stored as bulk data and loaded by the editor on demand). This *won't* address the related problem that reading a large BulkData into a TArray is still not possible due to the 31-bit index limit on TArray. #jira UE-58325 #rb Gil.Gribb
[CL 5337393 by Richard TalbotWatkin in 4.22 branch]

+++ Engine/Source/Runtime/Engine/Private/LightMap.cpp
@@ -2509,7 +2509,7 @@ int32 TQuantizedLightSampleBulkData<QuantizedLightSampleType>::GetElementSize() 
  * @param ElementIndex	Element index to serialize
  */
 template<class QuantizedLightSampleType>
-void TQuantizedLightSampleBulkData<QuantizedLightSampleType>::SerializeElement( FArchive& Ar, void* Data, int32 ElementIndex )
+void TQuantizedLightSampleBulkData<QuantizedLightSampleType>::SerializeElement( FArchive& Ar, void* Data, int64 ElementIndex )
 {
 	QuantizedLightSampleType* QuantizedLightSample = (QuantizedLightSampleType*)Data + ElementIndex;
 	// serialize as colors
